Bye Swing
removed old classes (from Swing) and new DAP and ZVMZWM correction and adding Ferienplan

+++ src/main/java/net/aclrian/mpe/controller/FinishController.java
@@ -3,11 +3,8 @@ package net.aclrian.mpe.controller;
 import com.itextpdf.html2pdf.ConverterProperties;
 import com.itextpdf.html2pdf.HtmlConverter;
 import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.fxml.FXML;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
-import javafx.stage.Stage;
 import javafx.stage.Window;
 import net.aclrian.mpe.messdiener.Messdiener;
 import net.aclrian.mpe.messe.Messe;
@@ -19,7 +16,6 @@ import net.aclrian.mpe.utils.DateienVerwalter;
 import net.aclrian.mpe.utils.Dialogs;
 import net.aclrian.mpe.utils.Log;
 import net.aclrian.mpe.utils.RemoveDoppelte;
-import org.apache.log4j.BasicConfigurator;
 import org.docx4j.jaxb.Context;
 import org.docx4j.openpackaging.packages.WordprocessingMLPackage;
 import org.docx4j.openpackaging.parts.WordprocessingML.NumberingDefinitionsPart;
@@ -28,7 +24,6 @@ import org.docx4j.convert.in.xhtml.XHTMLImporterImpl;
 
 import java.awt.*;
 import java.io.*;
-import java.lang.reflect.Array;
 import java.nio.charset.StandardCharsets;
 import java.text.SimpleDateFormat;
 import java.util.*;
@@ -97,45 +92,6 @@ public class FinishController implements Controller {
 		return locked;
 	}
 
-	public void back() {
-		/*if (ferienplan) {
-
-			JOptionPane op = new JOptionPane(
-					"Soll die aktuelle Einteilung gel"+References.oe+"scht werden, um einen neuen Plan mit ggf. ge"+References.ae+"nderten Messen/Abmeldungen generieren zu k\"+References.oe+\"nnen?",
-					JOptionPane.QUESTION_MESSAGE, JOptionPane.YES_NO_CANCEL_OPTION);
-			JFrame f = new JFrame();
-			f.setAlwaysOnTop(true);
-			JDialog dialog = op.createDialog(f, "Frage");
-			WEinFrame.farbe(dialog);
-			dialog.setVisible(true);
-			try {
-				if ((int) op.getValue() != 2) {
-					zurueck((int) op.getValue() == 0, EnumActivePanel.Abmelden, ap);
-				}
-			} catch (Exception e2) {
-				Utilities.logging(getClass(), "toback:ActionListener",
-						"Die Eingabe wurde vom Benutzer abgebrochen.");
-			}
-		} else {
-			JOptionPane op = new JOptionPane(
-					"Soll die aktuelle Einteilung gel"+References.oe+"scht werden, um einen neuen Plan mit ggf. ge"+References.ae+"nderten Messen generieren zu k\"+References.oe+\"nnen?",
-					JOptionPane.QUESTION_MESSAGE, JOptionPane.YES_NO_CANCEL_OPTION);
-			JFrame f = new JFrame();
-			f.setAlwaysOnTop(true);
-			JDialog dialog = op.createDialog(f, "Frage");
-			WEinFrame.farbe(dialog);
-			dialog.setVisible(true);
-			try {
-				if ((int) op.getValue() != 2) {
-					zurueck((int) op.getValue() == 0, EnumActivePanel.Start, ap);
-				}
-			} catch (Exception e2) {
-				Utilities.logging(getClass(), "toback:ActionListener",
-						"Die Eingabe wurde vom Benutzer abgebrochen.");
-			}
-		}*/
-	}
-
 	public void zurueck(MainController mc) {
 		Boolean löschen = Dialogs.yesNoCancel("Ja", "Nein", "Hier bleiben", "Soll die aktuelle Einteilung gelöscht werden, um einen neuen Plan mit ggf. geänderten Messen generieren zu können?");
 		if (löschen == null) return;
@@ -163,7 +119,7 @@ public class FinishController implements Controller {
 		converterProperties.setCharset("UTF-8");
 		try {
 			File out = new File(Log.getWorkingDir().getAbsolutePath() + File.separator + titel + ".pdf");
-			out.delete();
+			if(out.exists())out.delete();
 			HtmlConverter.convertToPdf(new ByteArrayInputStream(editor.getHtmlText().replaceAll("<p></p>","<br>").getBytes(StandardCharsets.UTF_8)),
 					new FileOutputStream(out), converterProperties);
 			Desktop.getDesktop().open(out);
@@ -191,7 +147,7 @@ public class FinishController implements Controller {
 			wordMLPackage.getMainDocumentPart().getContent().addAll(XHTMLImporter.convert(input, Log.getWorkingDir().getAbsolutePath()));
 
 			File out = new File(Log.getWorkingDir()+File.separator +titel+".docx");
-			out.delete();
+			if(out.exists())out.delete();
 			wordMLPackage.save(out);
 			Desktop.getDesktop().open(out);
 		} catch(Exception e){
@@ -201,64 +157,6 @@ public class FinishController implements Controller {
 
 		public void neuerAlgorythmus() {
 		hauptarray.sort(Messdiener.einteilen);
-		// TODO EFI-Brist-ZIENTer und notwendigkeit? Dap und zvmzwm fruehzeitg erkennen und beheben
-		/*for (StandartMesse sm : DateienVerwalter.dv.getPfarrei().getStandardMessen()) {
-			int anz_real = 0;
-			int anz_monat = 0;
-
-			ArrayList<Messdiener> array = new ArrayList<>();
-			if (!(sm instanceof Sonstiges)) {
-				for (Messdiener medi : hauptarray) {
-					int id = 0;
-					if (medi.isIstLeiter()) {
-						id++;
-					}
-					int ii = DateienVerwalter.dv.getPfarrei().getSettings().getDaten(id).getAnz_dienen();
-					if (medi.getDienverhalten().getBestimmtes(sm) && ii != 0) {
-						array.add(medi);
-						anz_monat += medi.getMessdatenDaten().getkannnochAnz();
-					}
-				}
-				anz_real = anz_monat;
-				int benoetigte_anz = 0;
-				ArrayList<Messe> dfmessen = new ArrayList<Messe>();
-				Calendar start = Calendar.getInstance();
-				start.setTime(messen.get(0).getDate());
-				start.add(Calendar.MONTH, 1);
-				for (Messe me : messen) {
-					if (!(me.getStandardMesse() instanceof Sonstiges)) {
-						if (me.getDate().after(start.getTime())) {
-							anz_real = anz_real + anz_monat;
-							start.setTime(me.getDate());
-						}
-						dfmessen.add(me);
-						benoetigte_anz += sm.getAnz_messdiener();
-					}
-				}
-				if (benoetigte_anz > anz_real) {
-					int fehlt = benoetigte_anz - anz_real;
-					try {
-						int medishinzufuegen = (int) Math.ceil(fehlt / array.size());// abrunden
-						Utilities.logging(this.getClass(), "neuerAlorythmus",
-								"Es wurde abgerundet: " + (double) (fehlt / array.size()) + "-->" + medishinzufuegen);
-						for (Messdiener messdiener : array) {
-							messdiener.getMessdatenDaten().addtomaxanz(medishinzufuegen, messdiener.isIstLeiter());
-						}
-						String mitteilung = "Zu den Messdienern, die am " + sm.getWochentag() + " um " + sm.getBeginn_stunde()
-								+ " koennen, werden + " + medishinzufuegen
-								+ " zu ihrem normalen Wert hinzugef"+ References.ue +"gt!";
-						Utilities.logging(this.getClass(), "neuerAlorythmus",
-								mitteilung);
-						new Erroropener(new Exception(mitteilung));
-					} catch (ArithmeticException e) {
-						Log.getLogger().info("Kein Messdiener kann: "
-								+ sm.getWochentag() + sm.getBeginn_stunde() + ":" + sm.getBeginn_minute());
-					}
-				}
-			}
-		}*/
-		// DAV UND ZWMZVM ENDE
-		Log.getLogger().info("Überpruefung zu ende!");
 		// neuer Monat:
 		Calendar start = Calendar.getInstance();
 		start.setTime(messen.get(0).getDate());
@@ -267,7 +165,6 @@ public class FinishController implements Controller {
 		Log.getLogger().info("nächster Monat bei: " + df.format(start.getTime()));
 		// EIGENTLICHER ALGORYTHMUS
 			for (Messe me : messen) {
-				// while(m.isfertig() || stackover){
 				if (me.getDate().after(start.getTime())) {
 					start.add(Calendar.MONTH, 1);
 					Log.getLogger().info("nächster Monat: Es ist " + df.format(me.getDate()));
@@ -290,60 +187,67 @@ public class FinishController implements Controller {
 		case EinfachEinteilen:
 			ArrayList<Messdiener> medis;
 			boolean zwang = false;
-			// try {
-			medis = get(new Sonstiges(), m);
-			// } catch (Exception e) {
-			// Utilities.logging(this.getClass(), this.getClass().getEnclosingMethod(),
-			// e.getMessage());
-			// medis = beheben(m, ap.getAda());
-			// zwang = true;
-			// }
+			medis = get(new Sonstiges(), m, false, false);
 			Log.getLogger().info(medis.size() + " für " + m.getNochBenoetigte());
-			if (m.getNochBenoetigte() > medis.size()) {
-				Dialogs.error("Die Messe " + m.getID().replaceAll("\t","   ") + " hat zu wenige Messdiener. Noch " + m.getNochBenoetigte() + " werden benötigt. Es gibt aber nur "+ (medis.size()-1));
-			}
 			for (Messdiener medi : medis) {
-				einteilen(m, medi, zwang);
+				einteilen(m, medi, false, false);
 			}
+			if(!m.istFertig())zwang(m);
 			break;
 		case TypeBeachten:
 			ArrayList<Messdiener> medis2;
 			boolean zwang2 = false;
-			// try {
-			medis2 = get(m.getStandardMesse(), m);
-			// } catch (Exception e) {
-			// Utilities.logging(this.getClass(), this.getClass().getEnclosingMethod(),
-			// e.getMessage());
-			// medis2 = beheben(m, ap.getAda());
-			// zwang2 = true;
-			// }
+			medis2 = get(m.getStandardMesse(), m, false, false);
 			Log.getLogger().info(medis2.size() + " für " + m.getNochBenoetigte());
-			if (m.getNochBenoetigte() > medis2.size()) {
-				Dialogs.error("Die Messe " + m.getID().replaceAll("\t","   ") + "hat zu wenige Messdiener");
-			}
 			for (Messdiener messdiener : medis2) {
-				einteilen(m, messdiener, zwang2);
+				einteilen(m, messdiener, false, false);
 			}
-			break;
-		default:
+			if(!m.istFertig())zwang(m);
 			break;
 		}
-		}
+	}
 
-		public void einteilen(Messe m, Messdiener medi, boolean zwang) {
-		boolean d = false;
+	private void zwang(Messe m) {
+		if(!(m.getStandardMesse() instanceof Sonstiges)) {
+			if (Dialogs.frage("Die Messe " + m.getID().replaceAll("\t", "   ") + " hat zu wenige Messdiener.\nNoch " + m.getNochBenoetigte() + " werden benötigt.\nSollen Messdiener eingeteilt werden, die standartmäßig die Messe \n'" + m.getStandardMesse().tokurzerBenutzerfreundlichenString() + "' dienen können, aber deren Anzahl schon zu hoch ist?")) {
+				ArrayList<Messdiener> medis = get(m.getStandardMesse(), m, false, true);
+				Log.getLogger().warn(m+" einteilen ohne Anzahl beachten");
+				for (Messdiener medi : medis) {
+					einteilen(m, medi, false,true);
+				}
+			}
+			if (m.istFertig()) return;
+			SimpleDateFormat df = new SimpleDateFormat("dd.MM.yyy");
+			if (Dialogs.frage("Die Messe " + m.getID().replaceAll("\t", "   ") + " hat zu wenige Messdiener.\nNoch " + m.getNochBenoetigte() + " werden benötigt.\nSollen Messdiener eingeteilt werden, die an dem Tag \n'" + df.format(m.getDate()) + "' dienen können?")) {
+				ArrayList<Messdiener> medis = get(new Sonstiges(), m, false, false);
+				Log.getLogger().warn(m+" einteilen ohne Standardmesse beachten");
+				for (Messdiener medi : medis) {
+					einteilen(m, medi, false,false);
+				}
+			}
+			if (m.istFertig()) return;
+		}
+		if(Dialogs.frage("Die Messe " + m.getID().replaceAll("\t","   ") + " hat zu wenige Messdiener.\nNoch " + m.getNochBenoetigte() + " werden benötigt.\nSollen Messdiener zwangsweise eingeteilt werden?")){
+			Log.getLogger().warn(m+" einteilen ohne Standardmesse beachten");
+			if(m.istFertig()) return;
+			ArrayList<Messdiener> medis = get(new Sonstiges(), m, true, true);
+			for (Messdiener medi : medis) {
+				einteilen(m, medi, true, true);
+			}
+		}
+		zuwenige(m);
+	}
+
+	private void zuwenige(Messe m){
+		if(!m.istFertig())Dialogs.error("Die Messe" + m.getID().replaceAll("\t","   ")+" wird zu wenige Messdiener haben.");
+	}
+
+		public void einteilen(Messe m, Messdiener medi, boolean zwangdate, boolean zwanganz) {
+		boolean d;
 		if (m.istFertig()) {
 			return;
-		} else if (zwang) {
-			if (medi.getMessdatenDaten().kann(m.getDate(), zwang)) {
-				m.einteilenZwang(medi);
-				d = true;
-			}
-		} else {
-			if (medi.getMessdatenDaten().kann(m.getDate(), zwang)) {
-				m.einteilen(medi);
-				d = true;
-			}
+		} else{
+			d = m.einteilen(medi, zwangdate, zwanganz);
 		}
 		if (!m.istFertig() && d) {
 			ArrayList<Messdiener> anv = medi.getMessdatenDaten().getAnvertraute(DateienVerwalter.dv.getAlleMedisVomOrdnerAlsList());
@@ -354,9 +258,9 @@ public class FinishController implements Controller {
 				anv.sort(Messdiener.einteilen);
 				for (Messdiener messdiener : anv) {
 					boolean b = messdiener.getDienverhalten().getBestimmtes(m.getStandardMesse());
-					if (messdiener.getMessdatenDaten().kann(m.getDate(), zwang) && b) {
-						Log.getLogger().info(messdiener.makeId() + " dient mit " + medi.makeId());
-						einteilen(m, messdiener, zwang);
+					if (messdiener.getMessdatenDaten().kann(m.getDate(), zwangdate, zwanganz) && b) {
+						Log.getLogger().info(messdiener.makeId() + " dient mit " + medi.makeId()+"?");
+						einteilen(m, messdiener, zwangdate, zwanganz);
 					}
 				}
 			}
@@ -408,7 +312,7 @@ public class FinishController implements Controller {
 		return rtn;
 		}*/
 
-		public ArrayList<Messdiener> get(StandartMesse sm, Messe m) {
+		public ArrayList<Messdiener> get(StandartMesse sm, Messe m, boolean zwangdate, boolean zwanganz) {
 		ArrayList<Messdiener> al = new ArrayList<>();
 		for (Messdiener medi : hauptarray) {
 			int id = 0;
@@ -417,7 +321,7 @@ public class FinishController implements Controller {
 			}
 			int ii = DateienVerwalter.dv.getPfarrei().getSettings().getDaten(id).getAnz_dienen();
 			if (medi.getDienverhalten().getBestimmtes(sm) && ii != 0
-					&& medi.getMessdatenDaten().kann(m.getDate(), false)) {
+					&& medi.getMessdatenDaten().kann(m.getDate(), zwangdate, zwanganz)) {
 				al.add(medi);
 			}
 		}
@@ -425,8 +329,8 @@ public class FinishController implements Controller {
 		al.sort(Messdiener.einteilen);
 		return al;
 		}
-
-		public ArrayList<Messdiener> get(StandartMesse sm, Date d) {
+@Deprecated
+		public ArrayList<Messdiener> get(StandartMesse sm, Date d, boolean zwangdate, boolean zwanganz) {
 		ArrayList<Messdiener> al = new ArrayList<>();
 		ArrayList<Messdiener> al2 = new ArrayList<>();
 		for (Messdiener medi : hauptarray) {
@@ -436,9 +340,9 @@ public class FinishController implements Controller {
 			}
 			int ii = DateienVerwalter.dv.getPfarrei().getSettings().getDaten(id).getAnz_dienen();
 			if (medi.getDienverhalten().getBestimmtes(sm) && ii != 0) {
-				if (medi.getMessdatenDaten().kann(d, false)) {
+				if (medi.getMessdatenDaten().kann(d, zwangdate, zwanganz)) {
 					al.add(medi);
-				} else if (medi.getMessdatenDaten().kann(d, true)) {
+				} else if (medi.getMessdatenDaten().kann(d, true,true)) {
 					al2.add(medi);
 				}
 			}
